make file intergration added. And window management improved
`<leader> m` now runs make outside of vim instead of within vim. It
seems when running make as a vim command, when ever make would generate
an error, the current buffer would go to a temp make buffer, which is
annoying. So running make as a terminal command (`:!make`) solves that.

Also instead of pressing <C-W> to switch windows in vim, it has been
changed to just holding down control. And tmux window switching has been
changed to `prefix + control vim keys` to not confict with vim