Native neovim lsp set up, not used
Native lsp functions the same as coc. However there are still a few bugs
with it. Especially the error messages does not wrap, it just goes off
the screen, which is pretty annoying.

So this config has gone back to coc and does not load the native lsp
stuff. The config are still there. To go back to native lsp, uncomment the
native lsp section.

Also fzf.vim has been switched to telescope

And <leader>df has been mapped to `:diffthis` in vim