leader m keys changed from mapping to function
Instead of mapping leader m based on file type, which was pretty
unreliable. As if I'm in a cpp file, and I open a markdown file, the md
map would override the cpp map. So i have to do `:e` in cpp buffer for
the mapping to take place.

Now `<leader>m` is mapped to a function, which excutes different
commands based on file type.